refactor attributes to a concrete attribute set which can be filtered and reused across class/method/etc
